requirements.txt
Django
django-filter
djangorestframework
psycopg2-binary
python-dotenv
